feat: truncate short description exceeding the byte limit
